requirements.txt
# requirements.txt
openai-whisper
python-vlc
ffmpeg-python
numpy